drf_test/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
